+++ requests_cache/serializers/preconf.py
@@ -5,6 +5,7 @@
    :nosignatures:
 """
 
+import json
 import pickle
 from functools import partial
 from importlib import import_module
@@ -107,22 +108,44 @@ except ImportError as e:
     bson_document_serializer = get_placeholder_class(e)
 
 
-# JSON serializer
-try:
-    import ujson as json
-
-    _json_preconf_stage = ujson_preconf_stage
-except ImportError:
-    import json  # type: ignore
-
-    _json_preconf_stage = json_preconf_stage
-
-_json_stage = Stage(dumps=partial(json.dumps, indent=2), loads=json.loads)
+# JSON serializer: stlib
 json_serializer = SerializerPipeline(
-    [_json_preconf_stage, _json_stage],
+    [json_preconf_stage, Stage(dumps=partial(json.dumps, indent=2), loads=json.loads)],
     name='json',
     is_binary=False,
-)  #: Complete JSON serializer; uses ultrajson if available
+)  #: Complete JSON serializer using stdlib json module
+default_json_serializer = json_serializer  #: Complete JSON serializer; uses orjson or ultrajson if available, otherwise stdlib json  # noqa: E501
+
+
+# JSON serializer: ultrajson
+try:
+    import ujson
+
+    ujson_serializer = SerializerPipeline(
+        [ujson_preconf_stage, Stage(dumps=partial(ujson.dumps, indent=2), loads=ujson.loads)],
+        name='ujson',
+        is_binary=False,
+    )  #: Complete JSON serializer using ultrajson module
+    default_json_serializer = ujson_serializer
+except ImportError as e:
+    ujson_serializer = get_placeholder_class(e)
+
+
+# JSON serializer: orjson
+try:
+    import orjson
+
+    orjson_serializer = SerializerPipeline(
+        [
+            orjson_preconf_stage,
+            Stage(dumps=partial(orjson.dumps, option=orjson.OPT_INDENT_2), loads=orjson.loads),
+        ],
+        name='orjson',
+        is_binary=True,
+    )  #: Complete JSON serializer using orjson module
+    default_json_serializer = orjson_serializer
+except ImportError as e:
+    orjson_serializer = get_placeholder_class(e)
 
 
 # YAML serializer