+++ HISTORY.md
@@ -5,6 +5,9 @@
 * Ignore and log timezone errors when attempting to reuse responses cached in `requests-cache <= 1.1`
 * Fix error handling with `stale_if_error` during revalidation requests
 * Remove `[bson]` package extra to prevent accidentally installing it in the same environment as `pymongo`
+* ⚠️ Changed behavior for `serializer='json'`, depending on installed packages:
+  * Priority: `orjson` -> `ujson` -> stdlib `json`
+* Added the following serializers to explicitly use a specific JSON library: `json_serializer`, `ujson_serializer`, and `orjson_serializer`
 
 ## 1.2.1 (2024-06-18)